"shadowTrigger" command refactoring
