+++ src/commands/shadowTrigger/constants.js
@@ -1,0 +1,9 @@
+/* eslint-disable-next-line import/prefer-default-export */
+export const DEFAULT_OPTIONS = Object.freeze({
+  log: true,
+  force: false,
+  bubbles: true,
+  cancelable: true,
+  timeout: 4000,
+  composed: true,
+});
