+++ src/commands/shadowTrigger/index.js
@@ -1,0 +1,3 @@
+import command from './command';
+
+export default command;
